fix: move call fn to test's lib
