+++ e2e/cases/cli.tests.js
@@ -1,29 +1,12 @@
 import { existsSync } from 'node:fs';
 import { join } from 'node:path';
 
-import { spawn } from 'node:child_process';
-import process from 'node:process';
-
 import sizeOf from 'image-size';
 import { rimrafSync } from 'rimraf';
 
-import { getBaseUrl, fixedSize, fixedFile } from '../lib';
-
-async function callCmd(targetTest, additionalArgs) {
-  const npx = process.platform === 'win32' ? 'npx.cmd' : 'npx';
-  const browser = process.env.BROWSER != null ? process.env.BROWSER : 'chrome:headless';
-  const baseArgs = ['testcafe', browser, targetTest, '-s', 'path=e2e/screens', '-q'];
-  const args = baseArgs.concat(additionalArgs);
-
-  const result = await new Promise(function fn(resolve) {
-    const cmd = spawn(npx, args, { shell: true, stdio: 'inherit' });
-
-    cmd.on('close', function onClose() {
-      resolve();
-    });
-  });
-  return result;
-}
+import {
+  getBaseUrl, fixedSize, fixedFile, callTC,
+} from '../lib';
 
 fixture('Testing cli --take-snapshot').page(getBaseUrl());
 
@@ -31,7 +14,7 @@ test('should care arg without name', async t => {
   const basePath = join('e2e', 'screens', 'Testing_cli', 'assert__it');
   rimrafSync(basePath);
 
-  await callCmd('e2e/file/cases/takesnapshot.test.js', ['--take-snapshot', 'base', '-t', 'takesnapshot']);
+  await callTC('e2e/file/cases/takesnapshot.test.js', ['--take-snapshot', 'base', '-t', 'takesnapshot']);
 
   await t.expect(existsSync(join(basePath, fixedFile('base.png')))).ok();
 });
@@ -41,9 +24,9 @@ test('should care arg with name', async t => {
   rimrafSync(basePath);
 
   const fpath = 'e2e/file/cases/takesnapshot.test.js';
-  await callCmd(fpath, ['--take-snapshot', 'base', '-t', 'takesnapshot']);
-  await callCmd(fpath, ['--take-snapshot', 'actual', '-t', 'takesnapshot']);
-  await callCmd(fpath, ['--take-snapshot', 'foo', '-t', 'takesnapshot']);
+  await callTC(fpath, ['--take-snapshot', 'base', '-t', 'takesnapshot']);
+  await callTC(fpath, ['--take-snapshot', 'actual', '-t', 'takesnapshot']);
+  await callTC(fpath, ['--take-snapshot', 'foo', '-t', 'takesnapshot']);
 
   await t.expect(existsSync(join(basePath, fixedFile('base.png')))).ok();
   await t.expect(existsSync(join(basePath, fixedFile('actual.png')))).ok();
@@ -58,8 +41,8 @@ test('should care arg', async t => {
   rimrafSync(basePath);
 
   const fpath = 'e2e/file/cases/takesnapshot.test.js';
-  await callCmd(fpath, ['--take-snapshot', '--full-page', '-t', 'fullpage']);
-  await callCmd(fpath, ['--take-snapshot', '-t', 'nofullpage']);
+  await callTC(fpath, ['--take-snapshot', '--full-page', '-t', 'fullpage']);
+  await callTC(fpath, ['--take-snapshot', '-t', 'nofullpage']);
 
   const pngFull = sizeOf(join(basePath, 'assert__it__with__fullpage', fixedFile('base.png')));
   await t.expect(pngFull.width).eql(fixedSize(1290));
